+++ Content.Server/ADT/HideGhostWarp/HideGhostWarpComponent.cs
@@ -1,9 +1,9 @@
-namespace Content.Server.NoGhostFollow;
+namespace Content.Server.HideGhostWarp;
 
 /// <summary>
 ///  Вешаем на сущность, и к ней нельзя будет тепнуться через "Телепорт призрака".
 /// </summary>
 [RegisterComponent]
-public sealed partial class NoGhostFollowComponent : Component
+public sealed partial class HideGhostWarpComponent : Component
 {
 }
